Styling logged-in page, refactoring code for theme
