+++ src/theme.js
@@ -1,0 +1,33 @@
+import { createTheme } from '@mui/material/styles';
+
+const theme = createTheme({
+  palette: {
+    primary: {
+      light: '##039686',
+      main: '#017969',
+      dark: '##004d3e',
+      contrastText: '#fff'
+    },
+    secondary: {
+      light: '#b32126',
+      main: '#961020',
+      dark: '#790111',
+      contrastText: '#fff'
+    },
+    tertiary: {
+      light: '#29ab4c',
+      main: '#128a38',
+      dark: '#005a1a',
+      contrastText: '#fff'
+    },
+    status: {
+      danger: '#e53e3e'
+    },
+    neutral: {
+      main: '#64748B',
+      contrastText: '#fff'
+    }
+  }
+});
+
+export default theme;
